Consolidate test utils
