--- packages/history/__tests__/TestSequences/spyOn.js
@@ -1,15 +1,0 @@
-import mock from 'jest-mock';
-
-export default function spyOn(object, method) {
-  let original = object[method];
-  let spy = mock.fn();
-
-  object[method] = spy;
-
-  return {
-    spy,
-    destroy() {
-      object[method] = original;
-    }
-  };
-}
